requirements.txt
flask
feedparser
python-dateutil
deep-translator
